Android lint to NNBD
