+++ tools/android_lint/pubspec.yaml
@@ -4,7 +4,7 @@
 
 name: android_lint
 environment:
-  sdk: '>=2.12.0-0.0.dev <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 # Do not add any dependencies that require more than what is provided in
 # //third_party.pkg, //third_party/dart/pkg, or
